pyproject.toml
[project]
name = "langchain-app"
version = "0.1.0"
description = "A LangChain terminal application"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.1.0",
    "langchain-google-genai>=1.0.0",
    "langchain-community>=0.0.10",
    "typer>=0.12.0",
    "pydantic>=2.5.3",
    "python-dotenv>=1.0.0",
    "ebooklib>=0.18",
    "beautifulsoup4>=4.12.0",
    "PyMuPDF>=1.23.0",
    "Pillow>=10.0.0",
    "reportlab>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
langchain-app = "src.main:app"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["test"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=term-missing"
pythonpath = ["."]
filterwarnings = [
    "ignore::DeprecationWarning:importlib._bootstrap",
    "ignore::DeprecationWarning",
]